requirements.txt
# --- Web backend ---
fastapi==0.127.1
uvicorn==0.40.0
python-dotenv==1.0.1
python-multipart

# --- Auth (JWT + password hashing) ---
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
email-validator

# --- LangChain + Groq + Mongo memory ---
langchain==1.2.0
langchain-core==1.2.5
langchain-community==0.3.0
langchain-text-splitters==1.1.0
langchain-groq==0.2.0
langchain-mongodb==0.9.0
pymongo==4.15.5

# --- PDF loading ---
pypdf==5.0.1

# --- Embeddings + Vector store + Hybrid retrieval ---
sentence-transformers==3.1.1
langchain-huggingface==0.1.0
faiss-cpu==1.13.2
rank-bm25==0.2.2
bm25s==0.2.14

# --- Templates (serving index.html via Jinja2Templates) ---
Jinja2==3.1.6